Improved error handling in the JsonXmlEventReader.
The Json.createParser() method may throw an unchecked exception. I
find that to be a design mistake, but then I'm old fashioned and
always a bit slow to catch up on the new trends.

Anyway: the unchecked exception is now caught and converted into a
checked XMLStreamException, so that the JsonXmlInputFactory
implementation behaves correctly.