Added support for OSGi.
The JAR is now an actual bundle that exports the main package.

OSGi support is implemented in the exact same way as in Sprox.
Recently a new bundle plugin was released. I want to look in
migrating from the Felix bundle plugin to this new one. Once I do
I'll do that for Sprox first, then for this one.

http://wiki.osgi.org/wiki/Maven_Bundle_Plugin